♻️ refactor: split MediaTab into TabList and ViewModeToggle components
- Extracted TabList and ViewModeToggle from MediaTab for better separation of concerns
- Created individual subcomponents to improve modularity and maintainability
